refactor: Move client and server script under script tab
