+++ frontend/src/builder.d.ts
@@ -66,7 +66,7 @@ declare type HashString = `#${string}`;
 declare type RGBString = `rgb(${number}, ${number}, ${number})`;
 
 declare type LeftSidebarTabOption = "Components" | "Layers";
-declare type RightSidebarTabOption = "Properties" | "Data" | "Settings";
+declare type RightSidebarTabOption = "Properties" | "Script" | "Settings";
 
 declare type BuilderMode = "select" | "text" | "container" | "image" | "html";
