+++ frontend/src/components/BuilderRightPanel.vue
@@ -12,10 +12,10 @@
 		<div
 			class="sticky top-0 z-[12] flex w-full border-gray-200 bg-white p-[2px] text-sm dark:border-zinc-800 dark:bg-zinc-900">
 			<button
-				v-for="tab of ['Properties', 'Data', 'Settings']"
+				v-for="tab of (['Properties', 'Script', 'Settings'] as RightSidebarTabOption[])"
 				:key="tab"
 				class="mx-3 flex-1 p-2"
-				@click="store.rightPanelActiveTab = tab as RightSidebarTabOption"
+				@click="store.rightPanelActiveTab = tab"
 				:class="{
 					'border-b-[1px] border-gray-900 dark:border-zinc-500 dark:text-zinc-300':
 						store.rightPanelActiveTab === tab,
@@ -25,9 +25,9 @@
 			</button>
 		</div>
 		<BlockProperties v-show="store.rightPanelActiveTab === 'Properties'" class="p-4" />
-		<PageData
+		<PageScript
 			class="p-4"
-			v-show="store.rightPanelActiveTab === 'Data'"
+			v-show="store.rightPanelActiveTab === 'Script'"
 			:key="store.selectedPage"
 			v-if="store.selectedPage && store.getActivePage()" />
 		<PageSettings
@@ -40,7 +40,7 @@
 <script setup lang="ts">
 import useStore from "@/store";
 import BlockProperties from "./BlockProperties.vue";
-import PageData from "./PageData.vue";
+import PageScript from "./PageScript.vue";
 import PageSettings from "./PageSettings.vue";
 import PanelResizer from "./PanelResizer.vue";
 const store = useStore();
