+++ frontend/src/components/PageScript.vue
@@ -1,265 +1,143 @@
 <template>
-	<div
-		class="absolute bottom-0 z-40 h-fit w-full border-t border-gray-200 bg-white p-3 dark:border-zinc-700 dark:bg-zinc-900">
-		<PanelResizer
-			side="top"
-			:dimension="store.builderLayout.scriptEditorHeight"
-			:minDimension="100"
-			:maxDimension="600"
-			@resize="store.builderLayout.scriptEditorHeight = $event"></PanelResizer>
-		<div
-			class="flex gap-3"
-			:style="{
-				height: store.builderLayout.scriptEditorHeight + 'px',
-			}">
-			<div class="flex flex-col gap-3">
-				<div class="pt-2 text-xs font-bold uppercase text-gray-600">Client Scripts</div>
-				<div class="flex w-56 flex-col gap-1">
-					<a
-						v-for="script in attachedScriptResource.data"
-						href="#"
-						:class="{
-							'text-gray-600 dark:text-gray-300': activeScript !== script,
-							'font-medium dark:text-zinc-200': activeScript === script,
-						}"
-						@click="selectScript(script)"
-						class="group flex items-center justify-between gap-1 text-sm">
-						<div class="flex w-[90%] items-center gap-1">
-							<CSSIcon class="shrink-0" v-if="script.script_type === 'CSS'" />
-							<JavaScriptIcon class="shrink-0" v-if="script.script_type === 'JavaScript'" />
-							<span
-								class="truncate"
-								@blur="updateScriptName($event, script)"
-								@keydown.enter.stop.prevent="script.editable = false"
-								@dblclick="script.editable = true"
-								:contenteditable="script.editable">
-								{{ script.script_name }}
-							</span>
-						</div>
-						<FeatherIcon name="trash" class="h-3 w-3" @click.stop="deleteScript(script.name)"></FeatherIcon>
-					</a>
-					<div class="flex gap-2">
-						<Dropdown
-							:options="[
-								{ label: 'JavaScript', onClick: () => addScript('JavaScript') },
-								{ label: 'CSS', onClick: () => addScript('CSS') },
-							]"
-							size="sm"
-							placement="right">
-							<template v-slot="{ open }">
-								<Button
-									class="mt-2 w-full text-xs dark:bg-zinc-800 dark:text-zinc-200 dark:hover:bg-zinc-700"
-									@click="open">
-									New Script
-								</Button>
-							</template>
-						</Dropdown>
-						<Dropdown
-							v-if="clientScriptResource.data && clientScriptResource.data.length > 0"
-							:options="
-								clientScriptResource.data.map((d: {'name': string}) => {
-									return {
-										label: d.name,
-										onClick: () => {
-											attachScript(d.name);
-										},
-									};
-								})
-							"
-							size="sm"
-							placement="right">
-							<template v-slot="{ open }">
-								<Button
-									class="mt-2 w-full text-xs dark:bg-zinc-800 dark:text-zinc-200 dark:hover:bg-zinc-700"
-									@click="open">
-									Attach Script
-								</Button>
-							</template>
-						</Dropdown>
+	<div class="flex flex-col gap-4">
+		<div class="flex gap-2">
+			<Button
+				@click="showClientScriptEditor()"
+				class="flex-1 text-base dark:bg-zinc-800 dark:text-zinc-200 dark:hover:bg-zinc-700">
+				Client Script
+			</Button>
+			<Button
+				@click="showServerScriptEditor()"
+				class="flex-1 text-base dark:bg-zinc-800 dark:text-zinc-200 dark:hover:bg-zinc-700">
+				Data Script
+			</Button>
+		</div>
+		<CodeEditor v-model="store.pageData" type="JSON" label="Data Preview" :readonly="true"></CodeEditor>
+		<Dialog
+			style="z-index: 40"
+			class="overscroll-none"
+			:options="{
+				title: currentScriptEditor == 'data' ? 'Data Script' : 'Client Script',
+				size: '7xl',
+			}"
+			v-model="showDialog">
+			<template #body-content>
+				<div v-if="currentScriptEditor == 'client'">
+					<PageClientScriptManager :page="store.getActivePage()"></PageClientScriptManager>
+				</div>
+				<div v-else>
+					<CodeEditor
+						class="overscroll-none"
+						v-model="page.page_data_script"
+						type="Python"
+						height="60vh"
+						:show-line-numbers="true"></CodeEditor>
+					<span class="text-xs text-gray-600 dark:text-zinc-400">
+						Can be used to fetch dynamic data from server and pass it to the page.
+						<br />
+						eg. data.events = frappe.get_list("Event")
+					</span>
+					<div class="mt-2 flex flex-1">
+						<Button
+							@click="savePageDataScript"
+							class="w-full text-base dark:bg-zinc-800 dark:text-zinc-200 dark:hover:bg-zinc-700">
+							Save
+						</Button>
 					</div>
 				</div>
-			</div>
-			<div
-				class="flex h-full w-full items-center justify-center rounded bg-gray-100 text-base text-gray-600 dark:bg-zinc-800 dark:text-zinc-500"
-				v-show="!activeScript">
-				Select Script
-			</div>
-			<div v-if="activeScript" class="flex h-full w-full flex-col">
-				<span class="rounded-t-sm bg-gray-100 p-1 px-2 text-xs dark:bg-zinc-800 dark:text-zinc-100">
-					{{ activeScript.script_name }}
-				</span>
-				<CodeEditor
-					v-if="activeScript.script_type === 'JavaScript'"
-					:modelValue="activeScript.script"
-					@update:modelValue="updateScript"
-					type="JavaScript"
-					class="flex-1"
-					height="auto"
-					:show-line-numbers="true"></CodeEditor>
-				<CodeEditor
-					v-if="activeScript.script_type === 'CSS'"
-					:modelValue="activeScript.script"
-					@update:modelValue="updateScript"
-					type="CSS"
-					class="flex-1"
-					height="auto"
-					:show-line-numbers="true"></CodeEditor>
-			</div>
-		</div>
+			</template>
+		</Dialog>
 	</div>
 </template>
-<script setup lang="ts">
+<script lang="ts" setup>
+import { webPages } from "@/data/webPage";
 import useStore from "@/store";
 import { BuilderPage } from "@/types/Builder/BuilderPage";
-import { Dropdown, createListResource, createResource } from "frappe-ui";
-import { PropType, ref, watch } from "vue";
+import { Dialog } from "frappe-ui";
+import { onMounted, ref } from "vue";
 import CodeEditor from "./CodeEditor.vue";
-import CSSIcon from "./Icons/CSS.vue";
-import JavaScriptIcon from "./Icons/JavaScript.vue";
-import PanelResizer from "./PanelResizer.vue";
-
-type attachedScript = {
-	script: string;
-	script_type: string;
-	name: string;
-	script_name: string;
-	editable: boolean;
-};
+import PageClientScriptManager from "./PageClientScriptManager.vue";
 const store = useStore();
-const activeScript = ref<attachedScript | null>(null);
+const showDialog = ref(false);
+const page = ref<BuilderPage>(store.getActivePage());
 
-const props = defineProps({
-	page: {
-		type: Object as PropType<BuilderPage>,
-		required: true,
-	},
+onMounted(() => {
+	page.value = store.getActivePage();
 });
 
-const attachedScriptResource = createListResource({
-	doctype: "Builder Page Client Script",
-	parent: "Builder Page",
-	filters: {
-		parent: props.page.name,
-	},
-	fields: [
-		"builder_script.script",
-		"builder_script.script_type",
-		"builder_script.name as script_name",
-		"name",
-	],
-	orderBy: "`tabBuilder Page Client Script`.creation asc",
-	auto: true,
-});
+const currentScriptEditor = ref<"client" | "data">("client");
 
-const clientScriptResource = createListResource({
-	doctype: "Builder Client Script",
-	fields: ["script", "script_type", "name"],
-	auto: true,
-});
-
-const selectScript = (script: attachedScript) => {
-	activeScript.value = script;
-};
-
-const updateScript = (value: string) => {
-	if (!activeScript.value) return;
-	clientScriptResource.setValue
+const savePageDataScript = () => {
+	webPages.setValue
 		.submit({
-			name: activeScript.value.script_name,
-			script: value,
+			name: page.value.name,
+			page_data_script: page.value.page_data_script,
 		})
 		.then(() => {
-			attachedScriptResource.reload();
+			showDialog.value = false;
+			store.setPageData();
 		});
 };
 
-const addScript = (scriptType: "JavaScript" | "CSS") => {
-	clientScriptResource.insert
-		.submit({
-			script_type: scriptType,
-			script: scriptType === "JavaScript" ? "// Write your script here\n" : "/* Write your CSS here */\n",
-		})
-		.then((res: { name: string; script_type: string; script: string }) => {
-			attachedScriptResource.insert
-				.submit({
-					parent: props.page.name,
-					parenttype: "Builder Page",
-					parentfield: "client_scripts",
-					builder_script: res.name,
-				})
-				.then(async () => {
-					await attachedScriptResource.reload();
-					attachedScriptResource.data?.forEach((script: attachedScript) => {
-						if (script.script_name === res.name) {
-							selectScript(script);
-						}
-					});
-				});
-		});
+const showClientScriptEditor = () => {
+	currentScriptEditor.value = "client";
+	showDialog.value = true;
 };
 
-const attachScript = (builder_script_name: string) => {
-	attachedScriptResource.insert
-		.submit({
-			parent: props.page.name,
-			parenttype: "Builder Page",
-			parentfield: "client_scripts",
-			builder_script: builder_script_name,
-		})
-		.then(async () => {
-			await attachedScriptResource.reload();
-			attachedScriptResource.data?.forEach((script: attachedScript) => {
-				if (script.script_name === builder_script_name) {
-					selectScript(script);
-				}
-			});
-		});
+const showServerScriptEditor = () => {
+	currentScriptEditor.value = "data";
+	showDialog.value = true;
 };
-
-const deleteScript = (scriptName: string) => {
-	activeScript.value = null;
-	attachedScriptResource.delete.submit(scriptName).then(() => {
-		attachedScriptResource.reload();
-	});
-};
-
-const updateScriptName = (ev: Event, script: attachedScript) => {
-	const target = ev.target as HTMLElement;
-	const newName = target.innerText.trim();
-	createResource({
-		url: "frappe.client.rename_doc",
-	})
-		.submit({
-			doctype: "Builder Client Script",
-			old_name: script?.script_name,
-			new_name: newName,
-		})
-		.then(async () => {
-			await attachedScriptResource.reload();
-			await clientScriptResource.reload();
-			attachedScriptResource.data?.forEach((script: attachedScript) => {
-				if (script.script_name === newName) {
-					selectScript(script);
-				}
-			});
-		});
-};
-
-watch(
-	() => props.page,
-	async () => {
-		activeScript.value = null;
-		attachedScriptResource.filters.parent = props.page.name;
-		await attachedScriptResource.reload();
-		if (attachedScriptResource.data && attachedScriptResource.data.length > 0) {
-			selectScript(attachedScriptResource.data[0]);
-		}
-	}
-);
 </script>
-<style scoped>
-:deep(.editor > .ace_editor) {
-	border-top-left-radius: 0;
-	border-top-right-radius: 0;
+<style>
+[id^="headlessui-dialog-panel"] {
+	@apply dark:bg-zinc-800;
+}
+[id^="headlessui-dialog-panel"] > div {
+	@apply dark:bg-zinc-800;
+	@apply dark:text-zinc-50;
+}
+
+[id^="headlessui-dialog-panel"] header h3 {
+	@apply dark:text-white;
+}
+
+[id^="headlessui-dialog-panel"] button svg path {
+	@apply dark:fill-white;
+}
+[id^="headlessui-dialog-panel"] button {
+	@apply dark:text-white;
+	@apply dark:hover:bg-zinc-700;
+	@apply dark:bg-zinc-900;
+}
+[id^="headlessui-dialog-panel"] input {
+	@apply dark:bg-zinc-900;
+	@apply dark:border-zinc-800;
+}
+
+[id^="headlessui-dialog-panel"] input:focus {
+	@apply dark:ring-0;
+	@apply dark:border-zinc-700;
+}
+
+[id^="headlessui-dialog-panel"] input[type="checkbox"]:checked {
+	@apply dark:bg-zinc-700;
+}
+
+[id^="headlessui-dialog-panel"] input[type="checkbox"]:focus {
+	@apply dark:ring-zinc-700;
+	@apply dark:ring-offset-0;
+}
+
+[id^="headlessui-dialog-panel"] input[type="checkbox"]:hover {
+	@apply dark:bg-zinc-900;
+}
+
+[id^="headlessui-dialog-panel"] label > span {
+	@apply dark:text-gray-50;
+}
+
+[id^="headlessui-dialog"] [data-dialog] {
+	@apply dark:bg-black-overlay-800;
 }
 </style>
